+++ src/File/AbstractModelFile.php
@@ -68,6 +68,9 @@ abstract class AbstractModelFile extends AbstractFile
         );
     }
 
+    /**
+     * @throws \InvalidArgumentException
+     */
     public function writeFile(bool $withSrc = true): void
     {
         if (!$this->getModel()) {
@@ -105,32 +108,34 @@ abstract class AbstractModelFile extends AbstractFile
         return $this->model;
     }
 
-    public function getModelFromStructAttribute(StructAttributeModel $attribute = null): ?StructModel
+    public function getModelFromStructAttribute(?StructAttributeModel $attribute = null): ?StructModel
     {
         return $this->getStructAttribute($attribute)->getTypeStruct();
     }
 
-    public function getRestrictionFromStructAttribute(StructAttributeModel $attribute = null): ?StructModel
+    public function getRestrictionFromStructAttribute(?StructAttributeModel $attribute = null): ?StructModel
     {
         $model = $this->getModelFromStructAttribute($attribute);
-        if ($model instanceof StructModel) {
-            // list are mainly scalar values of basic types (string, int, etc.) or of Restriction values
-            if ($model->isList()) {
-                $subModel = $this->getModelByName($model->getList());
-                if ($subModel && $subModel->isRestriction()) {
-                    $model = $subModel;
-                } elseif (!$model->isRestriction()) {
-                    $model = null;
-                }
-            } elseif (!$model->isRestriction()) {
-                $model = null;
-            }
+
+        if (!$model) {
+            return null;
+        }
+
+        // lists are mainly scalar values of basic types (string, int, etc.) or of Restriction values
+        if ($model->isList()) {
+            $subModel = $this->getModelByName($model->getList());
+            $model = $subModel && $subModel->isRestriction() ? $subModel : (!$model->isRestriction() ? null : $model);
+        } elseif (!$model->isRestriction()) {
+            $model = null;
         }
 
         return $model;
     }
 
-    public function getStructAttributeType(StructAttributeModel $attribute = null, bool $namespaced = false, bool $returnArrayType = true): string
+    /**
+     * @throws \InvalidArgumentException
+     */
+    public function getStructAttributeType(?StructAttributeModel $attribute = null, bool $namespaced = false, bool $returnArrayType = true): string
     {
         $attribute = $this->getStructAttribute($attribute);
 
@@ -147,16 +152,12 @@ abstract class AbstractModelFile extends AbstractFile
 
         if (!empty($type) && ($struct = $this->getGenerator()->getStructByName($type))) {
             $inheritance = $struct->getTopInheritance();
-            if (!empty($inheritance)) {
-                $type = str_replace('[]', '', $inheritance);
-            } else {
-                $type = $struct->getPackagedName($namespaced);
-            }
+            $type = !empty($inheritance) ? str_replace('[]', '', $inheritance) : $struct->getPackagedName($namespaced);
         }
 
         $model = $this->getModelFromStructAttribute($attribute);
         if ($model instanceof StructModel) {
-            // issue #84: union is considered as string as it would be difficult to have a method that accepts multiple object types.
+            // issue #84: union is considered as string as it would be challenging to have a method that accepts multiple object types.
             // If the property has to be an object of multiple types => new issue...
             if ($model->isRestriction() || $model->isUnion()) {
                 $type = self::TYPE_STRING;
@@ -170,7 +171,10 @@ abstract class AbstractModelFile extends AbstractFile
         return $type;
     }
 
-    public function getStructAttributeTypeAsPhpType(StructAttributeModel $fromAttribute = null, bool $returnArrayType = true): string
+    /**
+     * @throws \InvalidArgumentException
+     */
+    public function getStructAttributeTypeAsPhpType(?StructAttributeModel $fromAttribute = null, bool $returnArrayType = true): string
     {
         $attribute = $this->getStructAttribute($fromAttribute);
 
@@ -191,7 +195,7 @@ abstract class AbstractModelFile extends AbstractFile
      * Also see http://www.w3schools.com/schema/schema_dtypes_numeric.asp.
      *
      * @param mixed $type
-     * @param null  $xsdTypesPath
+     * @param null $xsdTypesPath
      * @param mixed $fallback
      *
      * @return mixed
@@ -206,7 +210,7 @@ abstract class AbstractModelFile extends AbstractFile
      * Also see http://www.w3schools.com/schema/schema_dtypes_numeric.asp.
      *
      * @param mixed $type
-     * @param null  $xsdTypesPath
+     * @param null $xsdTypesPath
      * @param mixed $fallback
      *
      * @return mixed
@@ -281,7 +285,7 @@ abstract class AbstractModelFile extends AbstractFile
         return 'This class stands for %s %s';
     }
 
-    protected function defineModelAnnotationsFromWsdl(PhpAnnotationBlock $block, AbstractModel $model = null): self
+    protected function defineModelAnnotationsFromWsdl(PhpAnnotationBlock $block, ?AbstractModel $model = null): self
     {
         FileUtils::defineModelAnnotationsFromWsdl($block, $model instanceof AbstractModel ? $model : $this->getModel());
 
@@ -384,7 +388,7 @@ abstract class AbstractModelFile extends AbstractFile
 
     abstract protected function getMethodAnnotationBlock(PhpMethod $method): ?PhpAnnotationBlock;
 
-    protected function getStructAttribute(StructAttributeModel $attribute = null): ?StructAttributeModel
+    protected function getStructAttribute(?StructAttributeModel $attribute = null): ?StructAttributeModel
     {
         $struct = $this->getModel();
         if (empty($attribute) && $struct instanceof StructModel && 1 === $struct->getAttributes()->count()) {
@@ -394,28 +398,37 @@ abstract class AbstractModelFile extends AbstractFile
         return $attribute;
     }
 
-    protected function getStructAttributeTypeGetAnnotation(StructAttributeModel $attribute = null, bool $returnArrayType = true, bool $nullableItemType = false): string
+    protected function getStructAttributeTypeGetAnnotation(?StructAttributeModel $attribute = null, bool $returnArrayType = true, bool $nullableItemType = false): string
     {
         $attribute = $this->getStructAttribute($attribute);
 
         if ($attribute->isXml()) {
-            return '\\DOMDocument|string|null';
+            return '\DOMDocument|string|null';
         }
 
-        return sprintf('%s%s%s', $this->getStructAttributeTypeAsPhpType($attribute, false), $this->useBrackets($attribute, $returnArrayType) ? '[]' : '', !$nullableItemType && ($attribute->isRequired() || $attribute->isArray() || $attribute->isList()) ? '' : '|null');
+        return sprintf(
+            '%s%s%s',
+            $this->getStructAttributeTypeAsPhpType($attribute, false),
+            $this->useBrackets($attribute, $returnArrayType) ? '[]' : '',
+            !$nullableItemType && !$attribute->isNullable() && ($attribute->isRequired() || $attribute->isArray() || $attribute->isList()) ? '' : '|null'
+        );
     }
 
     protected function getStructAttributeTypeSetAnnotation(StructAttributeModel $attribute, bool $returnArrayType = true, bool $itemType = false): string
     {
         if ($attribute->isXml()) {
-            return '\\DOMDocument|string|null';
+            return '\DOMDocument|string|null';
         }
 
         if ($attribute->isList()) {
             return 'array|string';
         }
 
-        return sprintf('%s%s', $this->getStructAttributeTypeAsPhpType($attribute, $returnArrayType), $this->useBrackets($attribute, !$itemType) ? '[]' : '');
+        return sprintf(
+            '%s%s',
+            $this->getStructAttributeTypeAsPhpType($attribute, $returnArrayType),
+            $this->useBrackets($attribute, !$itemType) ? '[]' : ''
+        );
     }
 
     protected function useBrackets(StructAttributeModel $attribute, bool $returnArrayType = true): bool